requirements.txt
Django==1.7.5
Pillow==2.8.1
argparse==1.2.1
boto==2.37.0
dj-database-url==0.3.0
dj-static==0.0.6
django-ordered-model==0.4.1
django-toolbelt==0.0.1
gunicorn==19.3.0
psycopg2==2.6
static3==0.5.1
wsgiref==0.1.2
